perf(dnsserver,cache,upstream,recursive,querylog,tuning): derive hot-path pool/queue ceilings from tuned resources
Add tuning.ScaleByCores/ScaleByMemory and use them to replace 9 hardcoded
pool/queue/worker ceilings (UDP socket count, UDP workers per socket, the
background cache-writer semaphore, the ASN memo cap, the L2/querylog
write-queue caps, upstream/recursive connection pool sizes, DoH idle-conn
limits) so a big box can scale past today's fixed caps while a Raspberry Pi
keeps today's values via the same floor.

Also shard the L2 cache write-behind writer: previously one goroutine
draining one channel, now cores-scaled and keyed by question hash the same
way L1 already shards reads, removing the fixed ~12.8k writes/sec ceiling a
single writer imposed regardless of queue size. Adds
TestCloseFlushesAllWriterShards, proving Close() drains every shard before
returning rather than just one.

Verified with go build/vet/test -race/golangci-lint across the whole repo,
gated after every individual change, plus new BenchmarkCacheSetThroughput
and BenchmarkQueryLogRecord for the write paths that previously had no
throughput benchmark.

+++ internal/recursive/resolver_test.go
@@ -773,7 +773,7 @@ func TestResolverConnPoolKeyCap(t *testing.T) {
 		addr := "10.0.0.1:" + strconv.Itoa(i) // port varies so every key is distinct
 		r.putConn("udp", addr, &dns.Conn{Conn: c1})
 	}
-	if got := r.poolKeyCount.Load(); got != nsConnPoolMaxKeys {
+	if got := r.poolKeyCount.Load(); got != int64(nsConnPoolMaxKeys) {
 		t.Fatalf("poolKeyCount = %d, want exactly %d (capped)", got, nsConnPoolMaxKeys)
 	}
 }
